we did it

+++ challenge_day3/station6.py
@@ -1,3 +1,3 @@
 import math
 def solution_station_6(i):    
-    return round(math.sin(i), 4)
+    return float(round(math.sin(i), 4))
